+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 1.0.21
+
+### Patch Changes
+
+- package version bumps
+
 ## 1.0.20
 
 ### Patch Changes