+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 1.2.27
+
+### Patch Changes
+
+- update enum logic
+- Updated dependencies
+  - @forklaunch/validator@1.2.14
+  - @forklaunch/common@1.2.14
+  - @forklaunch/core@1.3.17
+  - @forklaunch/ws@1.2.27
+
 ## 1.2.26
 
 ### Patch Changes